refactor: CartRepository 및 ProductRepository 인터페이스로 수정
