+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.ui.cart
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.model.Product
 
 class CartViewModel(
